Move Paper entity database
- moved from academic db to pluto db
- added comment to Paper entity
- rename ord, org to order, organization
