+++ src/main/java/network/pluto/bibliotheca/repositories/CommentRepository.java
@@ -2,6 +2,7 @@ package network.pluto.bibliotheca.repositories;
 
 import network.pluto.bibliotheca.models.Comment;
 import network.pluto.bibliotheca.models.Member;
+import network.pluto.bibliotheca.models.Paper;
 import network.pluto.bibliotheca.models.Review;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -9,5 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
     Page<Comment> findByReview(Review review, Pageable pageable);
+    Page<Comment> findByPaper(Paper paper, Pageable pageable);
     long countByCreatedBy(Member createdBy);
+    long countByPaper(Paper paper);
 }
