+++ src/main/java/network/pluto/bibliotheca/repositories/PaperRepository.java
@@ -1,5 +1,6 @@
-package network.pluto.bibliotheca.academic;
+package network.pluto.bibliotheca.repositories;
 
+import network.pluto.bibliotheca.models.Paper;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
